chore: move coin fixtures
