+++ tests/conftest.py
@@ -1,8 +1,7 @@
 import pytest
-from brownie import ZERO_ADDRESS, Contract
-from brownie_tokens import ERC20, MintableForkToken
+from brownie import ZERO_ADDRESS
 
-pytest_plugins = ["fixtures.accounts", "fixtures.deployments"]
+pytest_plugins = ["fixtures.accounts", "fixtures.coins", "fixtures.deployments"]
 
 
 def pytest_addoption(parser):
@@ -26,91 +25,9 @@ def isolation_setup(fn_isolation):
     pass
 
 
-@pytest.fixture(scope="module")
-def rebase_coin(alice, ATokenMock, AaveLendingPoolMock, ERC20Mock):
-    aave_lending_pool_mock = AaveLendingPoolMock.deploy({"from": alice})
-    erc20_mock = ERC20Mock.deploy("ERC20MOCK", "ERC", 18, {"from": alice})
-    yield ATokenMock.deploy(
-        "Rebase Mock", "MOK", 18, erc20_mock, aave_lending_pool_mock, {"from": alice}
-    )
-
-
-@pytest.fixture(scope="module")
-def zap(DepositZapUSD, alice):
-    yield DepositZapUSD.deploy({"from": alice})
-
-
-@pytest.fixture(scope="module")
-def base_pool():
-    pool = Contract("0xbEbc44782C7dB0a1A60Cb6fe97d0b483032FF1C7")
-
-    # ensure the base pool is balanced so our tests are deterministic
-    max_balance = max([pool.balances(0), pool.balances(1) * 10 ** 12, pool.balances(2) * 10 ** 12])
-    ideal_balances = [max_balance, max_balance // 10 ** 12, max_balance // 10 ** 12]
-    for i, amount in enumerate(ideal_balances):
-        balance = pool.balances(i)
-        if balance < amount:
-            MintableForkToken(pool.coins(i))._mint_for_testing(pool, amount - balance)
-    pool.donate_admin_fees({"from": pool.owner()})
-
-    yield pool
-
-
-@pytest.fixture(scope="module")
-def base_pool_btc(alice, fee_receiver, implementation_btc, factory, implementation_rebase_btc):
-    pool = Contract("0x7fC77b5c7614E1533320Ea6DDc2Eb61fa00A9714")
-    factory.add_base_pool(
-        pool,
-        fee_receiver,
-        [implementation_btc, implementation_rebase_btc] + [ZERO_ADDRESS] * 8,
-        {"from": alice},
-    )
-
-    yield pool
-
-
 @pytest.fixture(scope="module", params=[0, 1])
 def is_rebase(request):
     yield request.param
-
-
-@pytest.fixture(scope="module")
-def wrapped_coins(is_rebase, coin, rebase_coin, base_lp_token):
-    if is_rebase:
-        yield [rebase_coin, base_lp_token]
-    else:
-        yield [coin, base_lp_token]
-
-
-@pytest.fixture(scope="module")
-def underlying_coins(coin, is_rebase, rebase_coin):
-    BASE_COINS = [
-        "0x6B175474E89094C44Da98b954EedeAC495271d0F",  # DAI
-        "0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48",  # USDC
-        "0xdAC17F958D2ee523a2206206994597C13D831ec7",  # USDT
-    ]
-    if is_rebase:
-        yield [rebase_coin] + [MintableForkToken(i) for i in BASE_COINS]
-    else:
-        yield [coin] + [MintableForkToken(i) for i in BASE_COINS]
-
-
-@pytest.fixture(scope="module")
-def plain_coins():
-    yield [ERC20(decimals=7), ERC20(decimals=9), ZERO_ADDRESS, ZERO_ADDRESS]
-
-
-@pytest.fixture(scope="module")
-def base_lp_token():
-    yield MintableForkToken("0x6c3F90f043a72FA612cbac8115EE7e52BDe6E490")
-
-
-@pytest.fixture(scope="module")
-def coin(pytestconfig):
-    yield ERC20(
-        decimals=pytestconfig.getoption("decimals"),
-        success=eval(pytestconfig.getoption("return_value")),
-    )
 
 
 @pytest.fixture(scope="module")
@@ -208,16 +125,6 @@ def approve_zap(alice, bob, zap, swap, underlying_coins, initial_amounts_underly
 
 
 # Added for rebase tokens
-
-
-@pytest.fixture(scope="module")
-def wrapped_rebase_coins(rebase_coin, base_lp_token):
-    yield [rebase_coin, base_lp_token]
-
-
-@pytest.fixture(scope="module")
-def wrapped_rebase_decimals(wrapped_rebase_coins):
-    yield [i.decimals() for i in wrapped_rebase_coins]
 
 
 @pytest.fixture(scope="module")
